requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.4
crispy-bootstrap4==2025.6
crispy-bootstrap5==2025.6
Django==4.2.23
django-bootstrap-v5==1.0.11
django-bootstrap5==25.1
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-filter==25.1
djangorestframework==3.16.0
pytz==2025.2
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2

